backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Dependencies
torch>=2.2.0
torchaudio>=2.2.0
openai-whisper
transformers>=4.36.2
sentencepiece>=0.1.99
protobuf>=4.25.2
accelerate>=0.25.0
scipy>=1.11.4
numpy>=1.26.3
ffmpeg-python>=0.2.0

# TTS - Coqui TTS
coqui-tts>=0.22.0
